Add docstrings and perform minor clean up
