+++ abb_rws_client/include/abb_rws_client/rws_state_publisher_ros.hpp
@@ -40,12 +40,13 @@
 
 #pragma once
 
-#include <abb_rws_client/rws_client.hpp>
-
 #include <string>
 
 #include <rclcpp/rclcpp.hpp>
 #include <sensor_msgs/msg/joint_state.hpp>
+
+#include <abb_egm_rws_managers/rws_manager.h>
+#include <abb_egm_rws_managers/system_data_parser.h>
 
 #include <abb_rapid_sm_addin_msgs/msg/runtime_state.hpp>
 #include <abb_robot_msgs/msg/rapid_task_state.hpp>
@@ -53,21 +54,62 @@
 
 namespace abb_rws_client
 {
-class RWSStatePublisherROS : RWSClient
+class RWSStatePublisherROS
 {
 public:
+  /**
+   * \brief Creates a state publisher.
+   *
+   * \param node ROS2 node.
+   * \param robot_ip IP address for the robot controller's RWS server.
+   * \param robot_poty Port number for the robot controller's RWS server.
+   */
   RWSStatePublisherROS(const rclcpp::Node::SharedPtr& node, const std::string& robot_ip, unsigned short robot_port);
 
 private:
+  /**
+   * \brief Time callback for receving and publishing of system states from robot.
+   */
   void timer_callback();
 
+  rclcpp::Node::SharedPtr node_;
   rclcpp::TimerBase::SharedPtr timer_;
 
+  /**
+   * \brief Manager for handling RWS communication with the robot controller.
+   */
+  abb::robot::RWSManager rws_manager_;
+
+  /**
+   * \brief Description of the connected robot controller.
+   */
+  abb::robot::RobotControllerDescription robot_controller_description_;
+
+  /**
+   * \brief Publisher for joint states.
+   */
   rclcpp::Publisher<sensor_msgs::msg::JointState>::SharedPtr joint_state_pub_;
+
+  /**
+   * \brief Publisher for general system states.
+   */
   rclcpp::Publisher<abb_robot_msgs::msg::SystemState>::SharedPtr system_state_pub_;
+
+  /**
+   * \brief Publisher for RobotWare StateMachine Add-In runtime states.
+   *
+   * Note: Only used if the Add-In is present in the system.
+   */
   rclcpp::Publisher<abb_rapid_sm_addin_msgs::msg::RuntimeState>::SharedPtr runtime_state_pub_;
 
+  /**
+   * \brief Motion data for each mechanical unit defined in the robot controller.
+   */
   abb::robot::MotionData motion_data_;
+
+  /**
+   * \brief Data about the robot controller's system state.
+   */
   abb::robot::SystemStateData system_state_data_;
 };
